refactor/#15: Coupon prop 사용 방식 컨벤션
